refactor: cleanup gmaps
